requirements.txt
opencv-python
edge_impulse_linux
paho-mqtt
configparser
numpy>=1.19
PyAudio==0.2.11
psutil>=5.8.0
six==1.16.0
